pyproject.toml
[project]
name = "deepagents"
version = "0.0.3"
description = "General purpose 'deep agent' with sub-agent spawning, todo list capabilities, and mock file system. Built on LangGraph."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "langgraph>=0.2.6",
    "langchain-anthropic>=0.1.23",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain>=0.2.14",
    "python-dotenv>=1.1.1",
    "langfuse>=3.2.1",
    "tavily-python>=0.7.10",
    "poethepoet>=0.36.0",
    "langchain-qwq>=0.2.0",
    "openai>=1.99.1",
    "langchain-openai>=0.3.28",
    "langchain-mcp-adapters>=0.1.9"
]


[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["deepagents"]
[tool.setuptools.package-dir]
"deepagents" = "src/deepagents"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.poe.tasks]
cli = "uv run python main.py --trace"